[feat] #71 EditView(yoonjiho37)
- 스토리보드, 코드정돈
- 필요없는 파일 삭제
